Fix bug in range.
The comment here accurately notes that we may need to look at the stored range immediately prior to the start of the range we query. It does so by calling the `range` method on the underlying `BTreeMap` with a range from its beginning up to the start of our queried range and then getting its `.last()` element.

If such an element exists, its start is used as the start of the range that we will use for our return value (by way of a second call to `self.tree.range()` on it).

However, this does not handle the case where the last stored range beginning prior to the start of the queried range does not actually contain that starting point. If it doesn't, we still use the start of the prior range as the start of the range to return, which produces a range including elements that do not intersect our queried range.

Solve this by only extending the returned range backwards if the prior range really does include the start point of the queried range.

We factor this out into a helper, elem_range_that_may_overlap_range, to use in both range and range_mut.

remove is not impacted by this problem because it explicitly checks if the considered range contains the requested point.